added missing startup movie

+++ Data/FileHandler.cs
@@ -12,10 +12,12 @@ namespace AsylumLauncher
         private readonly string Legalus = Path.Combine(Directory.GetCurrentDirectory(), "..\\BmGame\\Movies\\Legalus.bik");
         private readonly string NvidiaVid = Path.Combine(Directory.GetCurrentDirectory(), "..\\BmGame\\Movies\\nvidia.bik");
         private readonly string UtLogo = Path.Combine(Directory.GetCurrentDirectory(), "..\\BmGame\\Movies\\UtLogo.bik");
+        private readonly string baalogo = Path.Combine(Directory.GetCurrentDirectory(), "..\\BmGame\\Movies\\baa_logo_run_v5_h264.bik");
         private readonly string LegalRenamed = Path.Combine(Directory.GetCurrentDirectory(), "..\\BmGame\\Movies\\Legal.bik.bak");
         private readonly string LegalusRenamed = Path.Combine(Directory.GetCurrentDirectory(), "..\\BmGame\\Movies\\Legalus.bik.bak");
         private readonly string NvidiaVidRenamed = Path.Combine(Directory.GetCurrentDirectory(), "..\\BmGame\\Movies\\nvidia.bik.bak");
         private readonly string UtLogoRenamed = Path.Combine(Directory.GetCurrentDirectory(), "..\\BmGame\\Movies\\UtLogo.bik.bak");
+        private readonly string baalogoRenamed = Path.Combine(Directory.GetCurrentDirectory(), "..\\BmGame\\Movies\\baa_logo_run_v5_h264.bik.bak");
         public readonly string ConfigDirectoryPath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments), "Square Enix\\Batman Arkham Asylum GOTY\\BmGame\\Config");
 
         public string BmEnginePath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments), "Square Enix\\Batman Arkham Asylum GOTY\\BmGame\\Config\\BmEngine.ini");
@@ -175,25 +177,26 @@ namespace AsylumLauncher
                 return;
             }
 
-            if (File.Exists(Legal) && File.Exists(Legalus) && File.Exists(NvidiaVid) && File.Exists(UtLogo) && File.Exists(LegalRenamed) && File.Exists(LegalusRenamed) && File.Exists(NvidiaVidRenamed) && File.Exists(UtLogoRenamed))
+            if (File.Exists(Legal) && File.Exists(Legalus) && File.Exists(NvidiaVid) && File.Exists(UtLogo) && File.Exists(LegalRenamed) && File.Exists(baalogo) && File.Exists(LegalusRenamed) && File.Exists(NvidiaVidRenamed) && File.Exists(UtLogoRenamed) && File.Exists(baalogoRenamed))
             {
                 File.Delete(LegalRenamed);
                 File.Delete(LegalusRenamed);
                 File.Delete(NvidiaVidRenamed);
                 File.Delete(UtLogoRenamed);
+                File.Delete(baalogoRenamed);
                 Program.MainWindow.SkipIntroBox.Enabled = true;
                 Program.MainWindow.SkipIntroBox.Checked = false;
                 IntroFilesRenamed = false;
             }
 
-            if (File.Exists(Legal) && File.Exists(Legalus) && File.Exists(NvidiaVid) && File.Exists(UtLogo))
+            if (File.Exists(Legal) && File.Exists(Legalus) && File.Exists(NvidiaVid) && File.Exists(UtLogo) && File.Exists(baalogo))
             {
                 Program.MainWindow.SkipIntroBox.Enabled = true;
                 Program.MainWindow.SkipIntroBox.Checked = false;
                 IntroFilesRenamed = false;
             }
 
-            if (File.Exists(LegalRenamed) && File.Exists(LegalusRenamed) && File.Exists(NvidiaVidRenamed) && File.Exists(UtLogoRenamed))
+            if (File.Exists(LegalRenamed) && File.Exists(LegalusRenamed) && File.Exists(NvidiaVidRenamed) && File.Exists(UtLogoRenamed) && File.Exists(baalogoRenamed))
             {
                 Program.MainWindow.SkipIntroBox.Enabled = true;
                 Program.MainWindow.SkipIntroBox.Checked = true;
@@ -214,6 +217,7 @@ namespace AsylumLauncher
                 File.Move(Legalus, LegalusRenamed);
                 File.Move(NvidiaVid, NvidiaVidRenamed);
                 File.Move(UtLogo, UtLogoRenamed);
+                File.Move(baalogo, baalogoRenamed);
                 IntroFilesRenamed = !IntroFilesRenamed;
                 Nlog.Info("RenameIntroVideoFiles - Disabling Startup Movies.");
             }
@@ -223,6 +227,7 @@ namespace AsylumLauncher
                 File.Move(LegalusRenamed, Legalus);
                 File.Move(NvidiaVidRenamed, NvidiaVid);
                 File.Move(UtLogoRenamed, UtLogo);
+                File.Move(baalogoRenamed, baalogo);
                 IntroFilesRenamed = !IntroFilesRenamed;
                 Nlog.Info("RenameIntroVideoFiles - Enabling Startup Movies.");
             }
